requirements.txt
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.3
boto3==1.12.46
botocore==1.15.46
chardet==3.0.4
coverage==5.1
Django==2.2.12
django-admin-sortable==2.2.3
django-appconf==1.0.4
django-bootstrap-breadcrumbs==0.9.2
django-braces==1.14.0
django-compressor==2.4
django-compressor-autoprefixer==0.1.0
django-contrib-comments==1.9.2
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filebrowser-no-grappelli==3.8.0
django-filter==2.2.0
django-inmemorystorage==0.1.1
django-json-ld==0.0.4
django-libsass==0.8
django-model-utils==4.0.0
django-newsletter==0.9
django-pipeline==1.7.0
django-robots==4.0
django-star-ratings==0.9.1
django-storages==1.9.1
django-taggit==1.2.0
django-taggit-labels==0.7.0
django-taggit-serializer==0.1.7
django-threadedcomments==1.2
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.3
importlib-metadata==1.6.0
isodate==0.6.0
jmespath==0.9.5
jsmin==2.2.2
ldif3==3.1.1
libsass==0.19.4
Markdown==3.2.1
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
pyenchant==3.0.1
pyparsing==2.4.7
pytest==5.4.2
python-card-me==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-memcached==1.59
python-memcached-stats==0.1
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
surlex==0.2.0
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.1
unicodecsv==0.14.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
